+++ eslint.config.mjs
@@ -1,0 +1,4 @@
+import config from '@sl-design-system/eslint-config';
+
+/** @type {import('eslint').Linter.FlatConfig[]} */
+export default [...config];