+++ packages/components/spinner/src/spinner.spec.ts
@@ -1,7 +1,7 @@
-import type { Spinner } from './spinner.js';
-import { expect, fixture, should } from '@open-wc/testing';
+import { expect, fixture } from '@open-wc/testing';
 import { html } from 'lit';
 import '../register.js';
+import { type Spinner } from './spinner.js';
 
 describe('sl-spinner', () => {
   let el: Spinner;
@@ -15,7 +15,7 @@ describe('sl-spinner', () => {
   });
 
   it('should have size medium and no variant by default', () => {
-    expect(el).to.have.attribute('size','md');
+    expect(el).to.have.attribute('size', 'md');
     expect(el).not.to.have.attribute('variant');
   });
 });